+++ CHANGELOG.md
@@ -4,8 +4,7 @@
 
 - added all demand, fuel mix, and supply methods for NYISO
 - add all demand methods for PJM
-- add get latest supply for SPP
-- add get latest supply for MISO
+- add get latest supply for SPP, MISO, and ISONE
 
 ## v0.3.0 - Aug 3, 2022